refactor Account model
