+++ iaso/models/__init__.py
@@ -1,3 +1,4 @@
+from .account import Account, AccountFeatureFlag
 from .base import *
 from .bulk_create_user_csv_file import BulkCreateUserCsvFile
 from .comment import CommentIaso
@@ -43,6 +44,7 @@ __all__ = [
     "SUPERSET",
     "TEXT",
     "Account",
+    "AccountFeatureFlag",
     "BulkCreateUserCsvFile",
     "CommentIaso",
     "DataSource",
